test: convert countblank flat tests.rs to proper tests/ directory structure
