+++ crates/core/src/eval/functions/statistical/countblank/tests/success.rs
@@ -1,5 +1,5 @@
-use super::*;
-use crate::types::{ErrorKind, Value};
+use super::super::countblank_fn;
+use crate::types::Value;
 
 #[test]
 fn empty_string_is_blank() {
@@ -63,44 +63,4 @@ fn array_no_blanks() {
     ]);
     let r = countblank_fn(&[arr]);
     assert_eq!(r, Value::Number(0.0));
-}
-
-#[test]
-fn no_args_returns_na() {
-    assert_eq!(countblank_fn(&[]), Value::Error(ErrorKind::NA));
-}
-
-#[test]
-fn array_with_empty_values_counted() {
-    // Array containing Value::Empty elements — should count as blank
-    let arr = Value::Array(vec![
-        Value::Empty,
-        Value::Text("a".to_string()),
-        Value::Empty,
-        Value::Text("b".to_string()),
-    ]);
-    assert_eq!(countblank_fn(&[arr]), Value::Number(2.0));
-}
-
-#[test]
-fn array_mixed_empty_and_blank_strings() {
-    // Both Value::Empty and empty Text are blank
-    let arr = Value::Array(vec![
-        Value::Empty,
-        Value::Text("".to_string()),
-        Value::Number(1.0),
-        Value::Text("hello".to_string()),
-    ]);
-    assert_eq!(countblank_fn(&[arr]), Value::Number(2.0));
-}
-
-#[test]
-fn array_with_no_blanks_returns_zero() {
-    // Array with non-blank values — count is 0
-    let arr = Value::Array(vec![
-        Value::Number(1.0),
-        Value::Text("a".to_string()),
-        Value::Bool(true),
-    ]);
-    assert_eq!(countblank_fn(&[arr]), Value::Number(0.0));
 }
